fix bogus handling of unreadable keys
Since the default action of the opendkim milter is to just log messages
as "can't determine message sende; accepting" when no key can be found
in the SigningTable, the correct action to a key that can't be read is
not to fail with a "file unreadable, message rejected" error but just
accept it and send it unsigned, as if there was no signing required in
the first place. This could be modified by the "relaxed" configuration
item, since it is really that setting that allows messages to be
accepted (and sent) unsigned if the sender is unknown.